モデル名修正
